requirements.txt

langchain
langchain-community
langchain-core
pydantic
langgraph
openai
streamlit
python-dotenv
tavily
python-dotenv 